requirements.txt
# opencv-python==4.8.1.78
# Pillow==10.1.0
# numpy==1.24.3
# tensorflow==2.14.0
# pyyaml==6.0.1
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# sqlalchemy==2.0.23
# pydantic==2.5.0
# requests==2.31.0
# psutil==5.9.6
# python-multipart==0.0.6
# python-dotenv==1.0.0
# matplotlib==3.8.2
opencv-python==4.10.0.84
numpy==1.26.4
tensorflow==2.18.0
Pillow==10.4.0
matplotlib==3.8.0
mediapipe==0.10.20
requests==2.32.3
python-dotenv==1.0.1
pyyaml==6.0.2
sqlalchemy==2.0.23
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
psutil==7.0.0